Refactor: move reused functions to logger_tools.py
